+++ food-delivery-backend/src/main/java/com/ait/app/service/MenuItemService.java
@@ -5,5 +5,5 @@ import com.ait.app.requestbody.MenuItemResponseDTO;
 
 public interface MenuItemService {
 
-	MenuItemResponseDTO addMenuItem(Long restaurantId,MenuItemRequestDTO request);
+	MenuItemResponseDTO addMenuItem(int  restaurantId,MenuItemRequestDTO request);
 }
